requirements.txt
google-api-python-client==2.108.0
youtube-transcript-api==0.6.1
pandas==2.1.3
python-dotenv==1.0.0 